firebase에로그인 관련 함수 생성 및 적용
